front: refactor agenda item modal (#373)
