+++ public/mobile-app/src/lib/components/modal/AgendaItemModal.svelte
@@ -1,94 +1,70 @@
 <script lang="ts">
-  import { onMount, type Snippet } from 'svelte';
+  import { type Agenda, buildAgenda, Item } from '$lib/agenda';
+  import ItemModal from '$lib/components/modal/ItemModal.svelte';
+  import { toastStore } from '$lib/state/toast.svelte';
 
   interface Props {
-    header: Snippet;
-    footer: Snippet;
-    onClose: () => void;
+    item: Item | null;
+    agenda: Agenda | null;
   }
-  let { header, footer, onClose }: Props = $props();
+  let { item = $bindable(), agenda = $bindable() }: Props = $props();
 
-  let dialog: HTMLDialogElement | null = null;
-
-  const handleClose = () => {
-    onClose();
+  const closeModal = () => {
+    item = null;
   };
 
-  onMount(() => {
-    if (dialog && !dialog.open) {
-      dialog.showModal();
-    }
-
-    dialog?.addEventListener('close', handleClose);
-
-    return () => {
-      dialog?.removeEventListener('close', handleClose);
-    };
-  });
-
-  const closeDialog = () => {
-    dialog?.close();
+  const refreshAgenda = () => {
+    buildAgenda().then((result) => {
+      agenda = result;
+    });
   };
 
-  const handleBackdropClick = (event: MouseEvent) => {
-    if (event.target === dialog) {
-      closeDialog();
+  const clickOnHideAgendaItem = (item: Item | null) => {
+    if (item) {
+      item.hide();
+      if (agenda) {
+        refreshAgenda();
+      }
+      closeModal();
+      toastStore.addToast("L'élément a bien été supprimé", 'success', 3000, true);
     }
   };
 </script>
 
-<dialog
-  bind:this={dialog}
-  class="modal"
-  onclick={handleBackdropClick}
-  data-testid="agenda-item-modal"
->
-  <div>
-    <button
-      onclick={closeDialog}
-      title="Fermer la modale"
-      class="close-button fr-icon-close-line"
-    ></button>
-    <div class="drag-handle"></div>
-    {@render header?.()}
-    {@render footer?.()}
-  </div>
-</dialog>
+<ItemModal onClose={closeModal}>
+  {#snippet header()}
+    <h2 class="agenda-item-modal-header" data-testid="agenda-item-modal-header">
+      {item?.title}
+    </h2>
+  {/snippet}
+
+  {#snippet footer()}
+    <ul class="agenda-item-modal-footer">
+      <li>
+        <span class="fr-icon-delete-line"></span>
+        <button
+          onclick={() => clickOnHideAgendaItem(item)}
+          title="Cacher l'élément de l'agenda"
+          aria-label="Cacher l'élément de l'agenda"
+          data-testid="hide-agenda-item-button"
+          class="hide-agenda-item"
+        >
+          Supprimer
+        </button>
+      </li>
+    </ul>
+  {/snippet}
+</ItemModal>
 
 <style>
-  dialog {
-    position: fixed;
-    top: auto;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: 0;
-    width: 100%;
-    max-width: 100%;
-    transform: none;
-
-    border-radius: 1.75rem 1.75rem 0 0;
-    border: none;
-    padding: 1rem 1rem 2rem;
-
-    .close-button {
-      position: absolute;
-      top: 1rem;
-      right: 1rem;
-      padding: 0;
-      color: var(--text-active-blue-france);
-    }
-
-    .drag-handle {
-      width: 2rem;
-      height: 0.25rem;
-      border-radius: 100px;
-      background-color: #79747e;
-      margin: 0 auto 1.75rem;
-    }
+  h2.agenda-item-modal-header {
+    font-size: 1.25rem;
   }
-  dialog::backdrop {
-    background: var(--grey-50-1000);
-    opacity: 64%;
+  ul.agenda-item-modal-footer {
+    padding: 0;
+    margin: 0;
+    li {
+      list-style: none;
+    }
   }
 </style>
