+++ public/mobile-app/src/routes/agenda/page.svelte.test.ts
@@ -215,7 +215,7 @@ describe('/+page.svelte', () => {
       });
 
       // Then
-      const agendaItemModal = screen.getByTestId('agenda-item-modal');
+      const agendaItemModal = screen.getByTestId('item-modal');
       expect(agendaItemModal).toBeInTheDocument();
     });
 
@@ -235,7 +235,7 @@ describe('/+page.svelte', () => {
         const moreIcon = screen.getByTestId('open-agenda-item-modal-fake-id-holiday-1');
         await fireEvent.click(moreIcon);
 
-        const agendaItemModal = screen.getByTestId('agenda-item-modal');
+        const agendaItemModal = screen.getByTestId('item-modal');
         expect(agendaItemModal).toBeInTheDocument();
 
         const deleteButton = screen.getByTestId('hide-agenda-item-button');
